refactor: isolate DatePicker stories
